- gestion de la vitesse en fonction du terrain et du ralentissement - ajustement de l'élasticité ______ - rajout pour ce soit plus beau - rajout d'animation, etc. ------ tous les effets ont ete rajoute ------ - ajout d'un retour de force lors de la frappe contre le palet dans toutes les directions, revision de la classe Controleur qui est maintenant Observable et PaletHaptique Observer ------ - ajout reussi d'une texture ------ - tentative ajout d'autre effet sans resultat - REamélioration de l'architecture logiciel ------ - tentative ajout d'autre effet sans resultat - REamélioration de l'architecture logiciel ------ - ajout de l'effet enclos ------ - gestion du score ------ - faire l'heritage sur le palet haptique, trouver les effets pour les autres evt haptique ! attention sur l'heritage du PaletHaptique il est aussi Observer ------ - préparation a l'ajout d'effet depuis un fichier ------ - interaction avec le palet, gestion des collisions mvc appliqué - ajout des buts - ajout de quelque detail terrain ------ - base du projet sans haptique - la raquette joueur se déplace

+++ trunk/ProjetVide/GestionCollision.cpp
@@ -19,7 +19,7 @@ bool GestionCollision::isCircleCollisionBetween(Palet *a, std::vector<Element *>
 	return false;
 }
 
-bool GestionCollision::isCircleGroundCollisionBetween(Palet *p, Terrain *t, int *dx, int *dy)
+bool GestionCollision::isCircleGroundCollisionBetween(Palet *p, Terrain *t, double *dx, double *dy)
 {
 	bool collision = false;
 	//détection a gauche du terrain
